Functional generator, updated ethereum and polygon interfaces

+++ indexer/types.go
@@ -44,7 +44,7 @@ type TransactionIndex struct {
 	Selector         string
 	TransactionHash  string // TODO: Rename this to Hash
 	TransactionIndex uint64 // TODO: Rename this to Index
-	Type             uint32
+	Type             uint64
 	Path             string
 }
 
@@ -53,7 +53,7 @@ func (t TransactionIndex) TableName() string {
 	return t.chain + "_transaction_index"
 }
 
-func NewTransactionIndex(chain string, blockNumber uint64, blockHash string, blockTimestamp uint64, fromAddress string, toAddress string, selector string, row_id uint64, transactionHash string, transactionIndex uint64, tx_type uint32, path string) TransactionIndex {
+func NewTransactionIndex(chain string, blockNumber uint64, blockHash string, blockTimestamp uint64, fromAddress string, toAddress string, selector string, row_id uint64, transactionHash string, transactionIndex uint64, tx_type uint64, path string) TransactionIndex {
 	return TransactionIndex{
 		chain:            chain,
 		BlockNumber:      blockNumber,
